fix all controller bugs

+++ utils/cloudinary.js
@@ -1,8 +1,6 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 import dotenv from 'dotenv'
-import { error } from "console";
-import exp from "constants";
 
 dotenv.config();
 
@@ -11,23 +9,39 @@ cloudinary.config({
   api_key: process.env.CLOUDINARY_API_KEY,
   api_secret: process.env.CLOUDINARY_API_SECRET, // Click 'View Credentials' below to copy your API secret
 });
+console.log({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY}
+)
 
 const uploadOnCloudinary = async (localFilePath) => {
   try {
     if (!localFilePath) return null;
+
+    if(!fs.existsSync(localFilePath)){
+      throw new Error("file not exists on local file path"+ localFilePath)
+    }
     //upload the file on cloudinary
 
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto",
     });
 
+    fs.unlinkSync(localFilePath);
+    
     /// file has been uploaded successfully.
-    console.log("file is uploaded on cloudinary", response.url);
+    // console.log("file is uploaded on cloudinary", response.url);
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // remove the locally saved file as the upload operation got failed
 
+    console.log("error uploading file", error.message || error)
+
+    if(fs.existsSync(localFilePath)){
+      fs.unlinkSync(localFilePath);
+    }
+     // remove the locally saved file as the upload operation got failed
+    return null;
   }
 };
 
-export {uploadOnCloudinary};
+export {uploadOnCloudinary};
